+++ inbm/Changelog.md
@@ -28,7 +28,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
  - (507873) Added support for Ubuntu 21.10
  - (507914) Add script to deploy .debs to a PPA hosted on Artifactory
  - (508492, 508493) [INBM] Support INBC Config commands(Get,Set)
-
+ - (508494) [INBM] Support INBC Config Load command
+ 
 ### Changed
  - (508706) - Change dynamic telemetry tag from ContainersCpuPercentage to ContainerStats